+++ usp/web_client/abstract_client.py
@@ -1,7 +1,9 @@
 """Abstract web client class."""
 
 import abc
+import random
 from http import HTTPStatus
+import time
 from typing import Optional
 
 RETRYABLE_HTTP_STATUS_CODES = {
@@ -187,3 +189,36 @@ class LocalWebClient(AbstractWebClient):
 
     def get(self, url: str) -> AbstractWebClientResponse:
         raise NoWebClientException
+
+
+class RequestWaiter:
+    """
+    Manages waiting between requests.
+    """
+
+    def __init__(self, wait: Optional[float] = None, random_wait: bool = True):
+        """
+        :param wait: time to wait between requests, in seconds.
+        :param random_wait: if true, wait time is multiplied by a random number between 0.5 and 1.5.
+        """
+        self.wait_s = wait or 0
+        self.random_wait = random_wait
+        self.is_first = True
+
+    def wait(self) -> None:
+        """Perform a wait if needed. Should be called before each request.
+
+        Will skip wait if this is the first request.
+        """
+        if self.wait_s == 0:
+            return
+
+        if self.is_first:
+            self.is_first = False
+            return
+
+        wait_f = 1.0
+        if self.random_wait:
+            wait_f = random.uniform(0.5, 1.5)
+
+        time.sleep(self.wait_s * wait_f)